# Remove all core-CH left-overs