--- a/unit_tests/library/options.cpp
+++ b/unit_tests/library/options.cpp
@@ -18,16 +18,6 @@ BOOST_AUTO_TEST_CASE(test_ch)
     OSRM osrm{config};
 }
 
-BOOST_AUTO_TEST_CASE(test_corech)
-{
-    using namespace osrm;
-    EngineConfig config;
-    config.use_shared_memory = false;
-    config.storage_config = storage::StorageConfig(OSRM_TEST_DATA_DIR "/corech/monaco.osrm");
-    config.algorithm = EngineConfig::Algorithm::CoreCH;
-    OSRM osrm{config};
-}
-
 BOOST_AUTO_TEST_CASE(test_mld)
 {
     using namespace osrm;