dead and gone.

<!-- BENCHMARK_RESULTS_START -->
## Benchmark Results
| Benchmark | Base | PR |
|-----------|------|----|
| alias | aliased u32: 1168.03<br/>plain u32: 1161.15<br/>aliased double: 1158.49<br/>plain double: 1218.61 | aliased u32: 1174.78<br/>plain u32: 1153.8<br/>aliased double: 1191.41<br/>plain double: 1185.45 |
| json-render | String: 6.66542ms<br/>Stringstream: 9.39526ms<br/>Vector: 6.87183ms | String: 6.57911ms<br/>Stringstream: 8.74616ms<br/>Vector: 6.87892ms |
| match_ch | Default radius: <br/>4.4393ms/req at 82 coordinate<br/>0.0541378ms/coordinate<br/>Radius 5m: <br/>4.41571ms/req at 82 coordinate<br/>0.0538501ms/coordinate<br/>Radius 10m: <br/>15.0888ms/req at 82 coordinate<br/>0.18401ms/coordinate<br/>Radius 15m: <br/>36.9426ms/req at 82 coordinate<br/>0.45052ms/coordinate<br/>Radius 30m: <br/>315.851ms/req at 82 coordinate<br/>3.85184ms/coordinate | Default radius: <br/>4.44845ms/req at 82 coordinate<br/>0.0542494ms/coordinate<br/>Radius 5m: <br/>4.44249ms/req at 82 coordinate<br/>0.0541768ms/coordinate<br/>Radius 10m: <br/>15.1716ms/req at 82 coordinate<br/>0.18502ms/coordinate<br/>Radius 15m: <br/>37.0263ms/req at 82 coordinate<br/>0.45154ms/coordinate<br/>Radius 30m: <br/>315.993ms/req at 82 coordinate<br/>3.85357ms/coordinate |
| match_mld | Default radius: <br/>2.87124ms/req at 82 coordinate<br/>0.0350151ms/coordinate<br/>Radius 5m: <br/>3.09398ms/req at 82 coordinate<br/>0.0377315ms/coordinate<br/>Radius 10m: <br/>11.2419ms/req at 82 coordinate<br/>0.137096ms/coordinate<br/>Radius 15m: <br/>28.4778ms/req at 82 coordinate<br/>0.347291ms/coordinate<br/>Radius 30m: <br/>339.321ms/req at 82 coordinate<br/>4.13806ms/coordinate | Default radius: <br/>2.83748ms/req at 82 coordinate<br/>0.0346034ms/coordinate<br/>Radius 5m: <br/>2.82259ms/req at 82 coordinate<br/>0.0344218ms/coordinate<br/>Radius 10m: <br/>10.5043ms/req at 82 coordinate<br/>0.128101ms/coordinate<br/>Radius 15m: <br/>26.97ms/req at 82 coordinate<br/>0.328902ms/coordinate<br/>Radius 30m: <br/>309.066ms/req at 82 coordinate<br/>3.7691ms/coordinate |
| packedvector | random write:<br/>std::vector 11415.9 ms<br/>util::packed_vector 81414.4 ms<br/>slowdown: 7.13169<br/>random read:<br/>std::vector 11102 ms<br/>util::packed_vector 33787.6 ms<br/>slowdown: 3.04339 | random write:<br/>std::vector 11469.2 ms<br/>util::packed_vector 78145.7 ms<br/>slowdown: 6.81352<br/>random read:<br/>std::vector 11673 ms<br/>util::packed_vector 33760.5 ms<br/>slowdown: 2.89217 |
| route_ch | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>507.49ms<br/>0.50749ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>346.986ms<br/>0.346986ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>625.868ms<br/>0.625868ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>152.42ms<br/>0.15242ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>109.439ms<br/>0.109439ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>132.601ms<br/>0.132601ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>150.244ms<br/>0.150244ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>96.9701ms<br/>0.0969701ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>132.257ms<br/>0.132257ms/req | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>512.049ms<br/>0.512049ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>348.944ms<br/>0.348944ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>628.77ms<br/>0.62877ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>151.73ms<br/>0.15173ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>97.4619ms<br/>0.0974619ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>132.656ms<br/>0.132656ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>151.646ms<br/>0.151646ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>97.4439ms<br/>0.0974439ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>132.554ms<br/>0.132554ms/req |
| route_mld | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>657.387ms<br/>0.657387ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>445.842ms<br/>0.445842ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>830.296ms<br/>0.830296ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>304.796ms<br/>0.304796ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>167.449ms<br/>0.167449ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>321.139ms<br/>0.321139ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>303.194ms<br/>0.303194ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>164.558ms<br/>0.164558ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>316.206ms<br/>0.316206ms/req | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>645.155ms<br/>0.645155ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>442.714ms<br/>0.442714ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>816.333ms<br/>0.816333ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>281.445ms<br/>0.281445ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>163.351ms<br/>0.163351ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>294.396ms<br/>0.294396ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>274.436ms<br/>0.274436ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>163.559ms<br/>0.163559ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>292.844ms<br/>0.292844ms/req |
| rtree | 1 result:<br/>207.233ms ->  0.0207233 ms/query<br/>10 results:<br/>241.991ms ->  0.0241991 ms/query | 1 result:<br/>207.145ms ->  0.0207145 ms/query<br/>10 results:<br/>242.546ms ->  0.0242546 ms/query |
<!-- BENCHMARK_RESULTS_END -->